+++ src/aioway/instrs/layers/norms.py
@@ -2,9 +2,6 @@
 
 
 from torch import nn
-from torchrl.data import tensor_specs as tspecs
-
-from aioway.deductions import deduction_for
 
 from ..nn import NnInstr, instr_dcls
 
@@ -80,13 +77,3 @@ class InstanceNorm3d(_BaseNorm):
     "Applies Instance Normalization over a 5D input."
 
     NN = nn.InstanceNorm3d
-
-
-@deduction_for(nn.BatchNorm1d).register
-@deduction_for(nn.BatchNorm2d).register
-@deduction_for(nn.BatchNorm3d).register
-@deduction_for(nn.InstanceNorm1d).register
-@deduction_for(nn.InstanceNorm2d).register
-@deduction_for(nn.InstanceNorm3d).register
-def norm_deduct(self, input: tspecs.Unbounded):
-    return input
